add session to delete subproject

+++ src/main/java/org/example/alphaplanner/controller/SubProjectController.java
@@ -56,6 +56,7 @@ public class SubProjectController {
 
     @PostMapping("/delete")
     private String deleteSubProject(HttpServletRequest request, HttpSession session, @RequestParam("subId") int subId) {
+        if (isNotLoggedIn(session)) return "redirect:";
         int pId = subProjectService.getSubProject(subId).getProjectId();
         if (!authorizationService.authProjectManager((Integer) session.getAttribute("userId"), pId)) return "redirect:";
         subProjectService.deleteSubProject(subId);
